move assets folder in
