+++ izzet/assets/loaders/gltf/gltf.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/go-gl/mathgl/mgl32"
-	"github.com/kkevinchou/izzet/internal/assets/assetslog"
+	"github.com/kkevinchou/izzet/izzet/assets/assetslog"
 	"github.com/kkevinchou/kitolib/modelspec"
 	"github.com/qmuntal/gltf"
 	"github.com/qmuntal/gltf/modeler"
