+++ izzet/assets/loaders/gltf/gltf_test.go
